+++ CS 172 Final Project/CS 172 Final Project/upgrade.cpp
@@ -1,6 +1,7 @@
 #include "upgrade.h"
 
-Upgrade::Upgrade(long int upgradeCppmAdditive, double upgradeCppmMultiplicative, long int upgradeRiskFactorAdditive, double upgradeRiskFactorMultiplicative, long int upgradeSellPriceAdditive, double upgradeSellPriceMultiplicative, long int upgradeProductionCostAdditive, double upgradeProductionCostMultiplicative, string upgradeName, long int upgradeCost, string upgradeDescription)
+Upgrade::Upgrade(long long int upgradeCppmAdditive, double upgradeCppmMultiplicative, int upgradeRiskFactorAdditive, double upgradeRiskFactorMultiplicative, long long int upgradeSellPriceAdditive, 
+	double upgradeSellPriceMultiplicative, long long int upgradeProductionCostAdditive, double upgradeProductionCostMultiplicative, string upgradeName, long long int upgradeCost, string upgradeDescription)
 {
 	this->upgradeCppmAdditive = upgradeCppmAdditive;
 	this->upgradeCppmMultiplicative = upgradeCppmMultiplicative;
@@ -21,7 +22,7 @@ Upgrade::Upgrade(long int upgradeCppmAdditive, double upgradeCppmMultiplicative,
 	this->upgradeDescription = upgradeDescription;
 }
 
-long int Upgrade::activeUpgrade(Production * productionMethod)
+long long int Upgrade::activeUpgrade(Production * productionMethod)
 {
 	productionMethod->setCookiesProduced((long)(productionMethod->getCookiesProduced()*upgradeCppmMultiplicative) + upgradeCppmAdditive);
     
@@ -39,7 +40,7 @@ void Upgrade::displayUpgradeParameters()
 	cout << upgradeName << " $" << upgradeCost << ": " << upgradeDescription;
 }
 
-long int Upgrade::getUpgradeCost()
+long long int Upgrade::getUpgradeCost()
 {
 	return upgradeCost;
 }
