+++ CS 172 Final Project/CS 172 Final Project/upgrade.h
@@ -12,34 +12,34 @@ class Upgrade {
 
 private:
 	//upgrade modifying values
-	long int upgradeCppmAdditive;
+	long long int upgradeCppmAdditive;
 	double upgradeCppmMultiplicative;
 
-	long int upgradeRiskFactorAdditive;
+	int upgradeRiskFactorAdditive;
 	double upgradeRiskFactorMultiplicative;
 
-	long int upgradeSellPriceAdditive;
+	long long int upgradeSellPriceAdditive;
 	double upgradeSellPriceMultiplicative;
 
-	long int upgradeProductionCostAdditive;
+	long long int upgradeProductionCostAdditive;
 	double upgradeProductionCostMultiplicative;
 
 	//upgrade parameters
 	string upgradeName;
-	long int upgradeCost;
+	long long int upgradeCost;
 	string upgradeDescription;
 
 public:
 	//class definition
-	Upgrade(long int upgradeCppmAdditive, double upgradeCppmMultiplicative, long int upgradeRiskFactorAdditive, double upgradeRiskFactorMultiplicative, long int upgradeSellPriceAdditive, double upgradeSellPriceMultiplicative, long int upgradeProductionCostAdditive, double upgradeProductionCostMultiplicative, string upgradeName, long int upgradeCost, string upgradeDescription);
+	Upgrade(long long int upgradeCppmAdditive, double upgradeCppmMultiplicative, int upgradeRiskFactorAdditive, double upgradeRiskFactorMultiplicative, long long int upgradeSellPriceAdditive, double upgradeSellPriceMultiplicative, long long int upgradeProductionCostAdditive, double upgradeProductionCostMultiplicative, string upgradeName, long long int upgradeCost, string upgradeDescription);
 
 	//I Don't quite remember what this was supposed to do
-	long int activeUpgrade(Production* productionMethod);
+	long long int activeUpgrade(Production* productionMethod);
 	
 	//I'm not entirely sure what this is supposed to do
 	//If I remember correctly it was to display the upgrade parameters to the menu
 	void displayUpgradeParameters();
 
 	// Returns the cost of the upgrade.
-	long int getUpgradeCost();
+	long long int getUpgradeCost();
 };
